E2E CI Test for Operator Bundle
This change augments the e2e test suite to simulate the
operator's deployment using OLM. The setup consists of the
following components:

- A docker/distribution container registry, running in
  docker outside of any Kubernetes cluster.
- A KinD cluster that is configured to resolve image refs
  which use "localhost" as the image registry domain.
- Installing OLM on the KinD cluster.

Once set up, the operator and its associated OLM bundle are
built and pushed to the local container registry. Next,
an OLM catalog is built based on the catalog published in
operatorhub.io. The catalog is what allows OLM to find the
Tekton operator that Shipwright depends on, and is likewise
pushed to the local container registry.

Building the operator, bundle, and catalog with a fully on-cluster
registry is problematic for several reasons:

- Not all tools can push to the on-cluster registry in this fashion
- Mainfests need to be rewritten to reference the on-cluster DNS name
for the registry
- The catalog source needs to be pullable within the cluster.

The test runs as follows:

- Create a namespace to run the operator under test
- Create a CatalogSource using the catalog containing the
  operator under test.
- Create an OperatorGroup which allows AllNamespace operators
  to be installed in the given namespace.
- Create a Subscription to install the Shipwright operator
  and its associated Tekton operator.
- Verify that the shipwright operator deploys successfully.

Contributor documentation has also been updated so that
developers can run this process using make commands on their
Kubernetes cluster of choice.

See also:
- https://kind.sigs.k8s.io/docs/user/local-registry/
- https://olm.operatorframework.io/docs/tasks/creating-a-catalog/
- https://olm.operatorframework.io/docs/tasks/make-catalog-available-on-cluster/
- https://olm.operatorframework.io/docs/tasks/install-operator-with-olm/
- https://olm.operatorframework.io/docs/advanced-tasks/operator-scoping-with-operatorgroups/