Reconcile Tekton in Separate Package
This change moves the logic that reconciles the installation of Tekton
into a separate package (`pkg/tekton`). The intent is to isolate - to
the best extent possible - the controller logic into discrete packages
and functions that can be readily unit tested. Functions should accept
interfaces so that tests can provide mocks/fakes as replacements - in
particular allowing fake Kubernetes clients in tests. The goal of the
unit tests is to focus exclusively on inputs and outputs. Tests that
verify the controller's behavior over multiple reconcile iterations
should be done with envtest-driven integration tests or black-box "e2e"
testing.

In `pkg/tekton`, the main `ReconcileTekton` function ensures that Tekton
Pipelines has been installed and that the TaskRun custom resource exists.
If the TaskRun CRD is not found, then the reconcile function attempts to
create an instance of the Tekton Operator's TektonConfig object. Existing
logic that checks the version of the Tekton Operator has been moved over.

The existing unit tests for the Tekton reconciliation have been moved to
unit tests for `ReconcileTekton` and related functions. Most test checks
were retained, with a few exceptions:

- The scenario that checks if a TektonConfig object exists and an
  apiserver error is thrown reading the TaskRun CRD has been dropped.
  There are separate unit tests that ensure if such an error is raised,
  it propagates outward.
- For consistency, the reconcile function asks for a requeue if the
  TektonConfig version is not valid for any reason.