Fix signing of released image
In 0b91c10b, the CSV stored in git was altered to include the `ko://`
image reference. The real image reference is generated at build time,
and its "resolved" CSV is stored in an ignored directory. This
unfortunately broke the release GitHub action, which greps the in-tree
CSV fo images to sign.

This updates the release action to look at the "resolved" CSV in the
`_output/olm` directory when searching for images to sign. This also
updates cosign-installer to v2.3.0.