CVE-2023-44487: Mitigate HTTP/2 Rapid Reset
Update base golang version and grpc dependency to mitigate
CVE-2023-44487 (HTTP/2 Rapid Reset) denial of service attack. grpc is an
indirect dependency used in our Kubernetes client libraries. Overall
risk for the operator is low, since the operator's typical deployment
does not expose an HTTP server that is reachable over the internet (it
does deploy a Prometheus metrics server, but that should not be exposed
to the general internet).