Add Developer Guide
This adds a developer guide that serves as an entrypoint for new code
contributors. The operator project has existing documentation for local
and OLM development, however it is buried in a nested directory that is
not easy to find. Those documents also miss critical foundational
knowledge, such as dev prerequisites and an understanding of OLM.

The new guide includes links to the Operator SDK and kubebuilder
tutorials, which will help contributors learn fundamentals like CRDs,
controllers, and packaging operators with OLM. I have also included
recommended tools when more than one choice is available in the market
(ex: docker or podman for a container engine). Finally, I have tried to
clarify when framework clis like `operator-sdk` and `kubebuilder` should
be used, and when it is safe to "just code."

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>