Set minKubeVersion to 1.24
Update KinD to latest released version (v0.20.0) and use the supported
v1.24 k8s node version from the corresponding release note. The current
Tekton operator on OperatorHub does not support k8s 1.25 or higher.
This also sets the minKubeVersion to 1.24.0 on the operator CSV so the
version we test with and the min supported version are aligned.

See https://github.com/tektoncd/operator/issues/1493.