SHIP-0038: Run CI on Release Branches
Update our CI-oriented GitHub actions to run when commits merge in a
`release-v*` branch, or a pull request is opened against a `release-v*`
branch. With this change, future release branches will automatically
have CI checks enabled. This commit should be backported to enable CI
in a prior release branch.

This implements a portion of SHIP-0038.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>