Source release.yaml from GitHub release The `release.yaml` used as the base for deploying Shipwright Builds appears to have been manipulated by a YAML linter or other processor. There are significant differentes in YAML structure and formatting between the `kodata/release.yaml` in tree and the `release.yaml` published on GitHub for the respective release (v0.12.0). This makes it difficult to verify that the operator is deploying the right thing, especially when versions are bumped.
This replaces the `kodata/release.yaml` in tree with the `release.yaml`
published for v0.12.0. This will make it easy to verify that we are
deploying the right version of Shipwright Builds with the operator, as
well as support future workflows that let contributors update the
deployed operand from a nightly release.

Signed-off-by: Ayush Satyam <asatyam@redhat.com>