Add RBAC for Build Webhook
In the v0.12.0 `release.yaml`, the Build API webhook has a cluster role
that grants it '*' permission on many core resources. This is
technically different from the original RBAC deployed in the operator's
modified `release.yaml`, which itemized the supported verbs for the
webhook. This commit adds the '*' verb permission for these resources,
which are then needed to deploy all content in `release.yaml`. The OLM
bundle was regenerated so that the RBAC is granted to the operator via
the `ClusterServiceVersion` definition.

Note that the use of '*' permissions in Kubernetes RBAC is highly
discouraged, as it violates the principle of least privilege. A more
narrowly scoped RBAC for the build webhook should be provided in a
future release.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>