Modifications from Dependency Updates
Upgrading to k8s 1.27 required a similar update to the version
Envtest simulates. While testing this locally, I discovered that the
latest branch of the setup-envtest tool requires golang 1.22. Using
the older `release-0.17` branch bypasses this issue for now.

After the update, some of the tests in
`TestShipwrightBuildReconciler_Reconcile` fail permanently. These
were previously marked as "brittle" because they used separate k8s test
clientsets. These tests are now skipped, resulting in loss of test
coverage related to operand image overrides. A refactor is in order to
ensure we only use the controller-runtime clientsets. This should come
after the upgrade to operator-sdk, which will require a restructure of
the project.

The update also exposed a latent bug in our controller logic, related
to updating the ShipwrightBuild object status. We previously were not
requeing on error, which in production can happen for a wide variety
of reasons. This bug was in fact caught by a lint check that was
skipped; this bypass has been removed to ensure we don't regress in
the future.