Initial Release Instructions
This is an initial guide on how to release the Shipwright
operator. It provides guidance on how to create a release
candidate build with GitHub actions, and adds release
verification instructions to the OLM development guide.

Future updates to this guide will include instructions on
how to publish the new bundle to OperatorHub.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>