doc: Procedure to Publish Operator on OperatorHub
Add a procedure and script to publish an operator release to the
upstream k8s community operators OperatorHub. The script requires the
user to fork the upstream k8s operators repsository and afterwards
submit a pull request by hand. The script also supports releasing to
any catalog that uses the OperatorHub file structure.

In the future this procedure/script can be added to our release
automation so when an operator release is published, we initiate the
process to update the upstream OperatorHub catalog.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>