chore: Code and Tool Updates From Dep Bumps
Updated `main.go` and the ShipwrightBuild controller test code to use
latest controller-runtime APIs (update included breaking changes).

Leaving the rest of the toolchain as is led to panics when generating
manifests with controller-gen. Therefore decided to update the
remaining tools to align with k8s 1.30.z:

- controller-gen to v0.15.0
- setup-envtest to release-0.18
- kustomize to v5.5.0 (most recent version)

This in turn required manifests to be regenerated. Operator-sdk was
left intact, as this does not appear to break with these updates.

Likewise updated Makefile and CI GitHub action to use golang 1.22 and
k8s 1.30, to ensure the version of k8s the operator tests with matches
the shipwright-io/build project version. Documentation for local and
OLM development updated to use latest versions of tools/operators -
notably updating Tekton Opertor to v0.74.0.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>